[ALPS02902098] [mt6763]display: sync code from dev branch
[Detail] remove unused code
add bypass_pq path
fix MIPI read reg error
fix dump info error

[Solution]

MTK-Commit-Id: d923c4cb049c51b74a0fa455816f047d35f3cbe4

CR-Id: ALPS02902098
Feature: Display Driver (DSI Interface)

+++ drivers/misc/mediatek/video/mt6763/dispsys/ddp_reg.h
@@ -155,7 +155,7 @@
 
 #define DISP_REG_GET(reg32) __raw_readl((unsigned long *)(reg32))
 #define DISP_REG_GET_FIELD(field, reg32) \
-	REG_FLD_VAL_GET(field, __raw_readl((unsigned int *)(reg32)))
+	REG_FLD_VAL_GET(field, __raw_readl((unsigned long *)(reg32)))
 
 /* polling register until masked bit is 1 */
 #define DDP_REG_POLLING(reg32, mask) \
